SmartErgonomicAssistantforDeskWorkers_Group4/requirements.txt
RPi.GPIO
gspread
oauth2client
lcd-drivers
